+++ src/components/Scroll Reveal/index.jsx
@@ -3,19 +3,22 @@ import PropTypes from 'prop-types';
 
 import sr from './Scroll Reveal';
 
+import './styles.scss';
+
 class ScrollReveal extends React.Component {
   componentDidMount() {
     sr.reveal('.scroll-reveal', this.props.config);
   }
 
   render() {
-    const { children } = this.props;
+    const { className, children } = this.props;
 
-    return <div className="scroll-reveal">{children}</div>;
+    return <div className={`scroll-reveal ${className}`}>{children}</div>;
   }
 }
 
 ScrollReveal.propTypes = {
+  className: PropTypes.string,
   children: PropTypes.any.isRequired,
   config: PropTypes.any
 };