+++ src/index.js
@@ -10,6 +10,10 @@ import ScrollToTop from './components/Scroll To Top';
 // Redux Store
 import { configureStore } from './store';
 
+import { library } from '@fortawesome/fontawesome-svg-core';
+import { faHeart, faEnvelope } from '@fortawesome/free-solid-svg-icons';
+import { fab } from '@fortawesome/free-brands-svg-icons';
+
 import App from './scenes/App';
 
 import '../resources/scss/main.scss';
@@ -19,6 +23,8 @@ const history = createHistory();
 
 // Configures store
 const store = configureStore();
+
+library.add(faHeart, faEnvelope, fab);
 
 ReactDOM.render(
   <Router history={history}>