+++ src/routes.jsx
@@ -3,7 +3,7 @@ import { Switch, Route, Redirect } from 'react-router-dom';
 
 import Home from './scenes/Home';
 import Blog from './scenes/Blog';
-import BlogPost from './scenes/BlogPost';
+import BlogPost from './scenes/Blog Post';
 import Team from './scenes/Team';
 import About from './scenes/About';
 import TeamMemberProfile from './scenes/Team Member Profile';