+++ src/scenes/Home/components/Home Masthead/index.jsx
@@ -1,19 +1,47 @@
 import React from 'react';
+import Proptypes from 'prop-types';
+
+import { connect } from 'react-redux';
+import { loadHeaderImage } from '../../../../store/base/base';
 
 import './styles.scss';
+
+// Maps Redux dispatch actions to props
+const mapDispatchToProps = dispatch => {
+  return {
+    loadHeaderImage: photoID => {
+      dispatch(loadHeaderImage(photoID));
+    }
+  };
+};
+
+// Maps Redux state to props
+function mapStateToProps(state) {
+  return { base: state.base };
+}
 
 class HomeMasthead extends React.Component {
   // Changes navbar colour upon component mount
   componentDidMount() {
+    const { loadHeaderImage } = this.props;
+
+    // Dispatches headerImage()
+    loadHeaderImage('1J9XegkkvKY06uCkiAywge');
+
+    // Changes navbar colour upon component mount
+    document.getElementById('navbar').classList.remove('navbar-light');
     document.getElementById('navbar').classList.add('navbar-dark');
   }
 
   // Resets navbar colour upon component unmount
   componentWillUnmount() {
     document.getElementById('navbar').classList.remove('navbar-dark');
+    document.getElementById('navbar').classList.add('navbar-light');
   }
 
   render() {
+    const { headerImage } = this.props.base;
+
     return (
       <div className="home-masthead">
         <div className="masthead-content-wrapper hv-center">
@@ -25,8 +53,8 @@ class HomeMasthead extends React.Component {
         <div className="masthead-image-wrapper">
           <img
             className="masthead-image"
-            src="https://source.unsplash.com/QSLdySbceR8/1600x1024"
-            alt=""
+            src={headerImage}
+            alt="The Lilac Foundation"
           />
         </div>
       </div>
@@ -34,4 +62,12 @@ class HomeMasthead extends React.Component {
   }
 }
 
-export default HomeMasthead;
+HomeMasthead.propTypes = {
+  loadHeaderImage: Proptypes.func,
+  base: Proptypes.object
+};
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(HomeMasthead);