+++ src/scenes/Project Post/index.jsx
@@ -2,10 +2,13 @@ import React from 'react';
 import Proptypes from 'prop-types';
 import { connect } from 'react-redux';
 
+import { Link } from 'react-router-dom';
 import * as Markdown from 'react-markdown';
 import moment from 'moment';
 
 import { loadProjectPost } from '../../store/projects/projects';
+
+import ShareItem from '../../components/Share Item';
 
 import './styles.scss';
 
@@ -42,13 +45,14 @@ class ProjectPost extends React.Component {
 
   render() {
     const { projectPost } = this.props;
-    const { title, date, content, projectPictures } = projectPost;
+    const { title, date, content } = projectPost;
 
-    return <div className="container">
-        <div className="row">
-          <div className="col-12 col-md-6 project-post-sidebar">
-            <div className="project-post-content">
-              <h2 className="project-post-title">{title}</h2>
+    return (
+      <div className="container">
+        <div className="row justify-content-center">
+          <div className="col-12 col-md-7">
+            <div className="project-post-wrapper">
+              <h1 className="project-post-title">{title}</h1>
               <p className="project-post-date">
                 {moment(date).calendar(null, {
                   sameDay: '[Today]',
@@ -60,18 +64,28 @@ class ProjectPost extends React.Component {
               <Markdown className="markdown-content" source={content} />
             </div>
           </div>
-          <div className="col-12 col-md-6 project-images hv-center" id="main">
-            <div>
-              <Markdown className="markdown-content" source={projectPictures} />
+        </div>
+        <div className="row">
+          <div className="col-12 col-md-4 container text-center">
+            <div className="pt-4">
+              <ShareItem pageTitle={title} />
+            </div>
+            <div className="py-4">
+              <Link className="text-muted" to="/projects">
+                More Projects
+              </Link>
             </div>
           </div>
         </div>
-      </div>;
+      </div>
+    );
   }
 }
 
 ProjectPost.propTypes = {
-  projectPost: Proptypes.any
+  projectPost: Proptypes.any,
+  location: Proptypes.object,
+  loadProjectPost: Proptypes.func
 };
 
 export default connect(