+++ src/store/base/actions.js
@@ -2,13 +2,23 @@
 
 // Import Redux base types
 import {
+    LOAD_LILAC_LOGO_SUCCESS,
     LOAD_ABOUT_ITEMS_SUCCESS,
     LOAD_PLACEHOLDER_SUCCESS,
     LOAD_HEADER_IMAGE_SUCCESS,
     LOAD_SOCIALS_SUCCESS,
+    LOAD_FOOTER_BLURB_SUCCESS
 } from './types';
 
 // Action that returns successfully retreived socials
+export function loadLilacLogoSuccess(lilacLogo) {
+    return {
+        type: LOAD_LILAC_LOGO_SUCCESS,
+        lilacLogo
+    }
+}
+
+// Action that rsuccessfully eturns retreived socials
 export function loadAboutItemsSuccess(aboutItems) {
     return {
         type: LOAD_ABOUT_ITEMS_SUCCESS,
@@ -16,7 +26,7 @@ export function loadAboutItemsSuccess(aboutItems) {
     }
 }
 
-// Action that returns the successfully loaded header image
+// Action that successfully returns loaded header image
 export function loadPlaceholderImageSuccess(placeholderImage) {
     return {
         type: LOAD_PLACEHOLDER_SUCCESS,
@@ -24,7 +34,7 @@ export function loadPlaceholderImageSuccess(placeholderImage) {
     };
 }
 
-// Action that returns the successfully loaded header image
+// Action that successfully returns loaded header image
 export function loadHeaderImageSuccess(headerImage) {
     return {
         type: LOAD_HEADER_IMAGE_SUCCESS,
@@ -32,10 +42,18 @@ export function loadHeaderImageSuccess(headerImage) {
     };
 }
 
-// Action that returns successfully retreived socials
+// Action that successfully returns retreived socials
 export function loadSocialsSuccess(socials) {
     return {
         type: LOAD_SOCIALS_SUCCESS,
         socials
+    };
+}
+
+// Action that successfully returns retreived socials
+export function loadFooterBlurbSuccess(footerBlurb) {
+    return {
+        type: LOAD_FOOTER_BLURB_SUCCESS,
+        footerBlurb
     };
 }