+++ src/store/base/base.js
@@ -2,42 +2,40 @@ import * as contentful from 'contentful';
 import * as actions from './actions';
 
 // The Lilac Foundation Client
-const blogClient = contentful.createClient({
-    space: 's85q1dv8e3m5',
-    accessToken: 'e1f99556ca22933d807bbeba1f93e70d5ebaa2bd37d43e099d26f653e9ade466'
-});
-
-// The Lilac Foundation Client
 const contentClient = contentful.createClient({
     space: 'lviz1x4y4a6n',
     accessToken: '788382ad2aed44f757676fb78d448b35e5633b7c2f3ca1bb456eec2c587c5231'
 });
 
-// Loads header image with id
-export function loadPlaceholderImage(id) {
+// Loads liac logo
+export function loadLilacLogo() {
     return dispatch => {
-        return blogClient.getAsset(id).then((asset) => {
-            console.log(asset)
-            dispatch(actions.loadPlaceholderImageSuccess(asset.fields.file.url))
-        });
+        return contentClient.getAsset('11JMH7ZEbGUK20s0YcIWIU').then((asset) =>
+            dispatch(actions.loadLilacLogoSuccess(asset.fields.file.url))).catch(error => console.log(error));
+    }
+}
+
+// Loads header image with id
+export function loadPlaceholderImage() {
+    return dispatch => {
+        return contentClient.getAsset('49YZTDMPagEM0uymEwEIA2').then((asset) =>
+            dispatch(actions.loadPlaceholderImageSuccess(asset.fields.file.url))).catch(error => console.log(error));
     }
 }
 
 // Loads header image with id
 export function loadHeaderImage(id) {
     return dispatch => {
-        return blogClient.getAsset(id).then((asset) => {
-            dispatch(actions.loadHeaderImageSuccess(asset.fields.file.url))
-        });
+        return contentClient.getAsset(id).then((asset) =>
+            dispatch(actions.loadHeaderImageSuccess(asset.fields.file.url))).catch(error => console.log(error));
     }
 }
 
 // Loads social media
 export function loadSocialMedia(id) {
     return dispatch => {
-        return contentClient.getEntry(id).then((socials) => {
-            dispatch(actions.loadSocialsSuccess(socials.fields))
-        }).catch(error => console.log(error));
+        return contentClient.getEntry(id).then((socials) =>
+            dispatch(actions.loadSocialsSuccess(socials.fields))).catch(error => console.log(error));
     }
 }
 
@@ -53,6 +51,15 @@ export function loadAboutItems() {
             }) =>
             // Dispatches loading success action return blog
             dispatch(actions.loadAboutItemsSuccess(items))
+        ).catch(error => console.log(error));
+    }
+}
+
+// Loads footer blurb with id
+export function loadFooterBlurb(id) {
+    return dispatch => {
+        return contentClient.getEntry(id).then((blurb) =>
+            dispatch(actions.loadFooterBlurbSuccess(blurb.fields))
         ).catch(error => console.log(error));
     }
 }