+++ src/store/base/reducer.js
@@ -1,16 +1,26 @@
 // Base reducer
 
 import initialState from '../initialState';
+
 import {
+    LOAD_LILAC_LOGO_SUCCESS,
     LOAD_PLACEHOLDER_SUCCESS,
     LOAD_HEADER_IMAGE_SUCCESS,
     LOAD_SOCIALS_SUCCESS,
-    LOAD_ABOUT_ITEMS_SUCCESS
+    LOAD_ABOUT_ITEMS_SUCCESS,
+    LOAD_FOOTER_BLURB_SUCCESS
 } from './types';
 
 // Base app reducer
 export default function baseReducer(state = initialState.base, action) {
     switch (action.type) {
+
+        // Successful lilac logo load type
+        case LOAD_LILAC_LOGO_SUCCESS:
+            return {
+                ...state,
+                lilacLogo: action.lilacLogo
+            };
 
         // Successful placeholder image load type
         case LOAD_PLACEHOLDER_SUCCESS:
@@ -38,6 +48,13 @@ export default function baseReducer(state = initialState.base, action) {
             return {
                 ...state,
                 aboutItems: action.aboutItems
+            }
+
+        // Successful footer blurb load type
+        case LOAD_FOOTER_BLURB_SUCCESS:
+            return {
+                ...state,
+                footerBlurb: action.footerBlurb
             }
 
         // Default case