+++ src/store/base/types.js
@@ -1,5 +1,8 @@
 // Base types
 
+// Successfully loaded liac logo
+export const LOAD_LILAC_LOGO_SUCCESS = 'LOAD_LILAC_LOGO_SUCCESS';
+
 // Successfully loaded header image
 export const LOAD_PLACEHOLDER_SUCCESS = 'LOAD_PLACEHOLDER_SUCCESS';
 
@@ -11,3 +14,6 @@ export const LOAD_ABOUT_ITEMS_SUCCESS = 'LOAD_ABOUT_ITEMS_SUCCESS';
 
 // Successfully loaded socials
 export const LOAD_SOCIALS_SUCCESS = 'LOAD_SOCIALS_SUCCESS';
+
+// Successfully loaded footer blurb
+export const LOAD_FOOTER_BLURB_SUCCESS = 'LOAD_FOOTER_BLURB_SUCCESS';