+++ src/store/initialState.js
@@ -1,6 +1,8 @@
 // Initial Redux state
 export default {
     base: {
+        lilacLogo: '',
+        footerBlurb: [],
         socials: [],
         headerImage: '',
         placeholderImage: ''